created Spanish Deck and refactored deck and spanish deck to both be children of GeneralDeck
