+++ src/main/java/models/Deck.java
@@ -1,12 +1,10 @@
 package models;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Random;
+//import java.util.ArrayList;
+//import java.util.Collections;
+//import java.util.Random;
 
-public class Deck {
-
-    public java.util.List<Card> cards = new ArrayList<>();
+public class Deck extends GeneralDeck{
 
     public Deck() {
         for(int i = 2; i < 15; i++){
@@ -17,19 +15,19 @@ public class Deck {
         }
     }
 
-    public void shuffle() {
-        long seed = System.nanoTime();
-        Collections.shuffle(cards, new Random(seed));
-    }
-
-    public ArrayList<Card> dealFour() {
-        ArrayList<Card> four = new ArrayList<>();
-        for (int i = 0; i < 4; i++) {
-            if (cards.size() > 0) {
-                four.add(cards.get(i));
-                cards.remove(i);
-            }
-        }
-        return four;
-    }
+//    public void shuffle() {
+//        long seed = System.nanoTime();
+//        Collections.shuffle(cards, new Random(seed));
+//    }
+//
+//    public ArrayList<Card> dealFour() {
+//        ArrayList<Card> four = new ArrayList<>();
+//        for (int i = 0; i < 4; i++) {
+//            if (cards.size() > 0) {
+//                four.add(cards.get(i));
+//                cards.remove(i);
+//            }
+//        }
+//        return four;
+//    }
 }
