+++ src/main/java/models/Suit.java
@@ -3,6 +3,6 @@ package models;
 import java.io.Serializable;
 
 public enum Suit  implements Serializable {
-    Hearts, Spades, Diamonds, Clubs, Cups, Coins, Swords
+    Hearts, Spades, Diamonds, Clubs, Cups, Coins, Swords, Jokers
 }
